requirements.txt
python-telegram-bot==20.0
Flask
SQLAlchemy
psycopg2-binary